+++ src/windowButtonsPanel/windowButton/index.ts
@@ -1,4 +1,4 @@
-// Copyright 2024, 2025 Glen Reesor
+// Copyright 2025 Glen Reesor
 //
 // This file is part of HammerBar.
 //
@@ -15,24 +15,6 @@
 // You should have received a copy of the GNU General Public License along with
 // HammerBar. If not, see <https://www.gnu.org/licenses/>.
 
-import { buildWindowButtonsPanel } from './buildWindowButtonsPanel';
+import { buildWindowButton } from './buildWindowButton';
 
-export function getWindowButtonsPanelBuilder(args: {
-  screenId: number;
-  windowStatusUpdateInterval: number;
-  showWindowPreviewOnHover: boolean;
-}) {
-  const { screenId, windowStatusUpdateInterval, showWindowPreviewOnHover } =
-    args;
-
-  return (args: {
-    coords: { x: number; y: number };
-    dimensions: { height: number; width: number };
-  }) =>
-    buildWindowButtonsPanel({
-      screenId,
-      windowStatusUpdateInterval,
-      showWindowPreviewOnHover,
-      ...args,
-    });
-}
+export { buildWindowButton };