+++ src/windowButtonsPanel/windowButton/types.ts
@@ -1,4 +1,4 @@
-// Copyright 2024, 2025 Glen Reesor
+// Copyright 2025 Glen Reesor
 //
 // This file is part of HammerBar.
 //
@@ -15,24 +15,17 @@
 // You should have received a copy of the GNU General Public License along with
 // HammerBar. If not, see <https://www.gnu.org/licenses/>.
 
-import { buildWindowButtonsPanel } from './buildWindowButtonsPanel';
-
-export function getWindowButtonsPanelBuilder(args: {
-  screenId: number;
-  windowStatusUpdateInterval: number;
-  showWindowPreviewOnHover: boolean;
-}) {
-  const { screenId, windowStatusUpdateInterval, showWindowPreviewOnHover } =
-    args;
-
-  return (args: {
-    coords: { x: number; y: number };
-    dimensions: { height: number; width: number };
-  }) =>
-    buildWindowButtonsPanel({
-      screenId,
-      windowStatusUpdateInterval,
-      showWindowPreviewOnHover,
-      ...args,
-    });
-}
+export type State = {
+  canvases: {
+    mainCanvas: hs.canvas.CanvasType | undefined;
+    hoverCanvas: hs.canvas.CanvasType | undefined;
+  };
+  mouseButtonIsDown: boolean;
+  mouseIsInsideButton: boolean;
+  x: number;
+  width: number;
+  windowObject: hs.window.WindowType;
+  windowTitle: string;
+  windowSnapshot: hs.image.ImageType | undefined;
+  isMinimized: boolean;
+};